src/agentstack_worker/requirements.txt
# =============================================================================
# AgentStack Worker - Python Dependencies
# =============================================================================

# Celery & Queue
celery[redis]==5.3.6
redis==5.0.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.2.4

# AI / LLM
openai==1.12.0
litellm==1.30.0

# Document Processing
docling==2.15.0
docling-core==2.15.0

# Observability (Phoenix)
arize-phoenix[otel]==4.40.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp==1.22.0
openinference-instrumentation-openai==0.1.6

# Utilities
python-dotenv==1.0.1
pydantic==2.6.0
httpx==0.26.0
tenacity==8.2.3
structlog==24.1.0